+++ src/main/java/frc/robot/subsystems/Drive/Swerve.java
@@ -32,10 +32,6 @@ import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import frc.lib.util.RumbleManager;
 import frc.robot.Constants;
-import frc.robot.SwerveConstants;
-import frc.robot.SwerveModule;
-// import frc.robot.Constants.frc.robot.config.RobotConfig.SWERVECONFIG;
-// import frc.robot.Constants.VisionConstants;
 import frc.robot.config.AutoConfig;
 import frc.robot.subsystems.Drive.GyroIO.GyroIOInputs;
 import java.util.HashMap;
